Work around xmllint (libxml2) "Relax-NG parser error" (bsc#1172131)
The choice construct in 'attribute ( t | type | config:type ) { "foo" }' is
correct Relax NG and processed by jing, but xmllint considers it incorrect
with "Found anyName attribute without oneOrMore ancestor".

So let's use slight duplication to work around it.

As a bonus it gets rid of the repeated namespace declarations in the RNG form.